+++ mry/src/mockable.rs
@@ -1,12 +1,12 @@
 #[diagnostic::on_unimplemented(
     message = "`{Self}` is not mockable argument because it is not `Send + 'static`",
-    note = "Consider `#[mry::mry(non_send(Rc, YourNotSendType))]` to enable SendWrapper, `#[mry::mry(skip_args(Rc, YourNotSendType))]` to skip the argument, or `#[mry::mry(skip_methods(method_name))]` to skip the method"
+    note = "Consider `#[mry::mry(non_send(Rc, YourNotSendType))]` to enable SendWrapper, `#[mry::mry(skip_args(Rc, YourNotSendType))]` to skip the argument, or `#[mry::mry(skip_fns(method_name))]` to skip the method"
 )]
 pub trait MockableArg: Send + 'static {}
 
 #[diagnostic::on_unimplemented(
     message = "`{Self}` is not mockable output because it is not `Send + 'static`",
-    note = "Consider `#[mry::mry(non_send(Rc, YourNotSendType))]` to enable SendWrapper, `#[mry::mry(skip_args(Rc, YourNotSendType))]` to skip the argument, or `#[mry::mry(skip_methods(method_name))]` to skip the method"
+    note = "Consider `#[mry::mry(non_send(Rc, YourNotSendType))]` to enable SendWrapper, `#[mry::mry(skip_args(Rc, YourNotSendType))]` to skip the argument, or `#[mry::mry(skip_fns(method_name))]` to skip the method"
 )]
 pub trait MockableRet: Send + 'static {}