+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.13
 
 - Added `#[mry::mry(skip_args(A, B, ...))]` attribute to skip types.
-- Added `#[mry::mry(skip_methods(method_name))]` attribute to skip some methods in trait or impl block.
+- Added `#[mry::mry(skip_fns(method_name))]` attribute to skip some methods in trait or impl block.
 
 ## 0.12